- created _setTagRights() - fixed search() + searchCount() signatures 0003266: error when removing all rights in shared tag edit dialog http://www.tine20.org/bugtracker/view.php?id=3266
git-svn-id: file:///www/metaways/svn.tine20.org/repo/trunk@16915 8d11cdf3-9d2f-0410-bbb5-e098f187f06b

+++ tine20/Admin/Controller/Tags.php
@@ -6,7 +6,7 @@
  * @subpackage  Controller
  * @license     http://www.gnu.org/licenses/agpl.html AGPL Version 3
  * @author      Lars Kneschke <l.kneschke@metaways.de>
- * @copyright   Copyright (c) 2007-2009 Metaways Infosystems GmbH (http://www.metaways.de)
+ * @copyright   Copyright (c) 2007-2010 Metaways Infosystems GmbH (http://www.metaways.de)
  * @version     $Id$
  * 
  * @todo        refactoring: use functions from Tinebase_Controller_Record_Abstract
@@ -60,24 +60,28 @@ class Admin_Controller_Tags extends Tinebase_Controller_Record_Abstract
     }
     
     /**
-     * get list of tags
+     * get list of records
      *
-     * @param Tinebase_Model_TagFilter $_filter
-     * @param Tinebase_Model_Pagination $_paging
-     * @return Tinebase_Record_RecordSet with record class Tinebase_Model_Tag
+     * @param Tinebase_Model_Filter_FilterGroup|optional $_filter
+     * @param Tinebase_Model_Pagination|optional $_pagination
+     * @param boolean $_getRelations
+     * @param boolean $_onlyIds
+     * @param string $_action for right/acl check
+     * @return Tinebase_Record_RecordSet|array
      */
-    public function search(Tinebase_Model_TagFilter $_filter, Tinebase_Model_Pagination $_paging)
+    public function search(Tinebase_Model_Filter_FilterGroup $_filter = NULL, Tinebase_Record_Interface $_pagination = NULL, $_getRelations = FALSE, $_onlyIds = FALSE, $_action = 'get')
     {
-        return Tinebase_Tags::getInstance()->searchTags($_filter, $_paging);
+        return Tinebase_Tags::getInstance()->searchTags($_filter, $_pagination);
     }
     
     /**
-     * get count of tags
-     *
-     * @param Tinebase_Model_TagFilter $_filter
+     * Gets total count of search with $_filter
+     * 
+     * @param Tinebase_Model_Filter_FilterGroup $_filter
+     * @param string $_action for right/acl check
      * @return int
      */
-    public function searchCount(Tinebase_Model_TagFilter $_filter)
+    public function searchCount(Tinebase_Model_Filter_FilterGroup $_filter, $_action = 'get')
     {
         return Tinebase_Tags::getInstance()->getSearchTagsCount($_filter);
     }
@@ -95,7 +99,7 @@ class Admin_Controller_Tags extends Tinebase_Controller_Record_Abstract
         $fullTag->rights =  Tinebase_Tags::getInstance()->getRights($_tagId);
         $fullTag->contexts = Tinebase_Tags::getInstance()->getContexts($_tagId);
         
-        return $fullTag;        
+        return $fullTag;
     }  
 
    /**
@@ -111,8 +115,7 @@ class Admin_Controller_Tags extends Tinebase_Controller_Record_Abstract
         $_tag->type = Tinebase_Model_Tag::TYPE_SHARED;
         $newTag = Tinebase_Tags::getInstance()->createTag(new Tinebase_Model_Tag($_tag->toArray(), true));
 
-        $_tag->rights->tag_id = $newTag->getId();
-        Tinebase_Tags::getInstance()->setRights($_tag->rights);
+        $this->_setTagRights($_tag, $newTag->getId());
         Tinebase_Tags::getInstance()->setContexts($_tag->contexts, $newTag->getId());
         
         return $this->get($newTag->getId());
@@ -130,15 +133,44 @@ class Admin_Controller_Tags extends Tinebase_Controller_Record_Abstract
         
         Tinebase_Tags::getInstance()->updateTag(new Tinebase_Model_Tag($_tag->toArray(), true));
         
-        $_tag->rights->tag_id = $_tag->getId();
-        Tinebase_Tags::getInstance()->purgeRights($_tag->getId());
-        Tinebase_Tags::getInstance()->setRights($_tag->rights);
-        
+        $this->_setTagRights($_tag, $_tag->getId(), TRUE);
         Tinebase_Tags::getInstance()->purgeContexts($_tag->getId());
         Tinebase_Tags::getInstance()->setContexts($_tag->contexts, $_tag->getId());
         
         return $this->get($_tag->getId());
     }  
+    
+    /**
+     * set tag rights
+     * 
+     * @param Tinebase_Model_FullTag $_tag
+     * @param string $_tagId
+     * @param boolean $_purgeRights
+     * @throws Tinebase_Exception_InvalidArgument
+     */
+    protected function _setTagRights(Tinebase_Model_FullTag $_tag, $_tagId, $_purgeRights = FALSE)
+    {
+        if (count($_tag->rights) == 0) {
+            throw new Tinebase_Exception_InvalidArgument('Could not save tag without rights');
+        } 
+        
+        // @todo tag needs at least 1 view_right to be usable
+        
+        
+//        $viewRightFound = FALSE;
+//        foreach ($_tag->rights as $right) {
+//            
+//        } 
+        
+        //&& $_tag->rights->getFirstRecord)
+        
+        $_tag->rights->tag_id = $_tagId;
+        
+        if ($_purgeRights) {
+            Tinebase_Tags::getInstance()->purgeRights($_tag->getId());
+        }
+        Tinebase_Tags::getInstance()->setRights($_tag->rights);        
+    }
     
     /**
      * delete multiple tags
@@ -152,5 +184,4 @@ class Admin_Controller_Tags extends Tinebase_Controller_Record_Abstract
         
         Tinebase_Tags::getInstance()->deleteTags($_tagIds);
     }
-
 }
